+++ jdaviz/core/loaders/tests/test_load_catalogs.py
@@ -374,6 +374,7 @@ def test_astroquery_load_catalog_source(deconfigged_helper):
     ldr.importer.col_other = ['parallax', 'pm', 'bp_rp', 'phot_rp_mean_mag']
     ldr.importer.viewer.create_new = 'Scatter'
     ldr.load()
+
     assert 'Scatter' in deconfigged_helper.viewers
     assert 'Catalog' in deconfigged_helper.viewers['Scatter'].data_menu.layer.choices